Updated the alert system (Challange)
I added a button to see all the collections from the data base. I implemented the alert system to alert the admin on the home page that there are problems with entries from the collection database. On the collection page, entries that have problems will be colored light red and with dark red I've colored the problem itself (the dumpster not existing in the "Pubela" database or the address either not existing in the "PubeleCetateni" database or not corresponding to the one in the database)